refactor(tests): split event stats/attendance-api tests into own file (Issue 1318)
test_event_stats.py mixed helper unit tests (_cancellations,
_resolve_cancelled_at, rsvp counts) with API endpoint tests
(TestStatsEndpoint, TestSetAttendance). Split the API endpoint
classes into test_event_attendance_api.py, which imports the
shared fixtures + _auth helper from test_event_stats.

No behavior change — pure move, done ahead of PR #1323 (which
would otherwise push this file over the 500-line hard cap).
